Move nested_documentable helper methods to a new common actions module
